refactor: relocate log helper to separate service
